UDP Regulations #20
